mqtt/requirements.txt
paho-mqtt
jsonpickle